fix(carousel): reorganize folder structure
