+++ src/components/organisms/Carousel/Carousel/Carousel.md
