+++ src/components/organisms/Carousel/CarouselSlide/CarouselSlide.md
