+++ src/components/organisms/Carousel/CarouselSlide/CarouselSlide.tsx
