+++ src/components/organisms/Carousel/CarouselSlideIcon/CarouselSlideIcon.md
