+++ src/components/organisms/Carousel/CarouselSlideIcon/CarouselSlideIcon.tsx
