+++ src/components/organisms/Carousel/CarouselSlideText/CarouselSlideText.md
