+++ src/components/organisms/Carousel/CarouselSlideText/CarouselSlideText.tsx
